First Upload
First Upload of the code I've been working on. Its basically the
BasicUse example from the TLC5940 Arduino Library. I've made a LED
Matrix Array with all of the pins, as numbered by the library, and the
letters A and B so far. Please feel Free to add to this code. I'm still
figuring out github so bear with me as I do.